S/W Mutation, Clock Tree defined, some self tuning
First, for all things the formatting of the communication or definition
has its own function like SetSPI_MasterHW_Format(....)
We keep in Timing() only what affect the clock frequencies.
The clock frequencies in the MCU are now in a global unique structure
pointed by each element.
Each frequency which can be tuned or modified is made of 3 elements: The
min, max and current real value.

If say the APB1 frequency is not defined, calling the Timing() function
will mostly update the min max range to narrow down the choices or find
a blocking point.

The clock tree includes the various PLLs.
Each time a frequency is divided or multiplied with a coefficient, it's
called a clock node.
We added the simplest solver: Based on the output desired min/max/best
desired frequencies, it is possible to run a full or step by step
itterative search or M,N,P,Q,R coefficient for any PLL. This resides in
ClockTree.*

We are also creating (not yet testing) the "gears", where we could have
up to 5 gears for the MCU core, keeping the peripheral APBx frequencies
fixed. More work to do there.
Also, we tried to support the OLED display... we can turn it on, invert
video but loading a picture in its video memory failed. Timed out on
this task.